+++ src/__tests__/index.spec.ts
@@ -5,6 +5,8 @@ describe("escape", () => {
   it("should escape everything", () => {
     t.equal(escape("a123<7&2\"'>"), "a123&lt;7&amp;2&quot;&#39;&gt;");
     t.equal(escape(1), "1");
+    t.equal(escape('"<>&'), "&quot;&lt;&gt;&amp;");
+    t.equal(escape("foo"), "foo");
     t.equal(escape(true), "true");
   });
 });