+++ src/index.ts
@@ -1,21 +1,54 @@
 export * from "./dom";
 
-const ESC: Record<string, string> = {
-  "<": "&lt;",
-  ">": "&gt;",
-  '"': "&quot;",
-  "'": "&#39;",
-  "&": "&amp;",
-};
+export function escape(text: string | number | boolean): string {
+  if (typeof text === "string") {
+    let result = text;
+    let escape: string = "";
+    let start = 0;
+    let i;
+    for (i = 0; i < text.length; i++) {
+      switch (text.charCodeAt(i)) {
+        case 34: // "
+          escape = "&quot;";
+          break;
+        case 39: // \
+          escape = "&#39;";
+          break;
+        case 38: // &
+          escape = "&amp;";
+          break;
+        case 60: // <
+          escape = "&lt;";
+          break;
+        case 62: // >
+          escape = "&gt;";
+          break;
+        default:
+          continue;
+      }
+      if (i > start) {
+        if (start > 0) {
+          result += text.slice(start, i);
+        } else {
+          result = text.slice(start, i);
+        }
+      }
 
-const escapeChar = (a: string) => ESC[a] || a;
+      if (i > 0) {
+        result += escape;
+      } else {
+        result = escape;
+      }
 
-export function escape(s: string | number | boolean): string {
-  if (typeof s === "string") {
-    return s.replace(/[<>"'&]/g, escapeChar);
+      start = i + 1;
+    }
+    if (start && i !== start) {
+      return result + text.slice(start, i);
+    }
+    return result;
+  } else {
+    return text.toString();
   }
-
-  return s.toString();
 }
 
 export const padCache = [
@@ -85,8 +118,9 @@ export let memoize = (fn: any, mem: any = {}) => (v: any) => {
   return mem[v];
 };
 
+const upperCaseRegex = /[A-Z]/g;
 // Taken from `preact-render-to-string` by @developit
 // See: https://github.com/developit/preact-render-to-string/blob/master/src/util.js#L65
 export let jsToCss = memoize((s: string) =>
-  s.replace(/([A-Z])/g, "-$1").toLowerCase(),
+  s.replace(upperCaseRegex, "-$&").toLowerCase(),
 );